Conflicts:
	src/protected/db-updates.php

+++ src/protected/application/lib/MapasCulturais/Entity.php
@@ -55,7 +55,7 @@ abstract class Entity implements \JsonSerializable{
     const STATUS_DISABLED = -9;
     const STATUS_TRASH = -10;
     const STATUS_ARCHIVED = -2;
-
+    
     /**
      * array of validation definition
      * @var array
@@ -91,16 +91,22 @@ abstract class Entity implements \JsonSerializable{
         if(property_exists($this, 'createTimestamp'))
                 $this->createTimestamp = new \DateTime;
 
-        if($this->usesTaxonomies())
-            $this->populateTermsProperty();
-
+        if($this->usesOwnerAgent() && !$app->user->is('guest')){
+            $this->setOwner($app->user->profile);
+        }
+        
         $hook_class_path = $this->getHookClassPath();
 
         App::i()->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').new');
+        
     }
 
     function __toString() {
         return $this->getClassName() . ':' . $this->id;
+    }
+    
+    static function isPrivateEntity(){
+        return false;
     }
 
     function refresh(){
@@ -239,6 +245,10 @@ abstract class Entity implements \JsonSerializable{
         }else{
             return $this->canUser('@control', $user);
         }
+    }
+    
+    protected function canUserModify($user) {
+        return $this->genericPermissionVerification($user);
     }
 
     protected function canUserRemove($user){
@@ -398,7 +408,7 @@ abstract class Entity implements \JsonSerializable{
 
         return $data_array;
     }
-    
+
     static public function getValidations(){
         return [];
     }
@@ -490,7 +500,7 @@ abstract class Entity implements \JsonSerializable{
     public function getEntityType(){
         return str_replace('MapasCulturais\Entities\\','',$this->getClassName());
     }
-    
+
     public function getEntityTypeLabel($plural = false) {}
 
     function getEntityState() {
@@ -524,6 +534,7 @@ abstract class Entity implements \JsonSerializable{
         if (method_exists($this, '_saveOwnerAgent')) {
             try {
                 $this->_saveOwnerAgent();
+                
             } catch (Exceptions\WorkflowRequestTransport $e) {
                 $requests[] = $e->request;
             }
@@ -532,7 +543,7 @@ abstract class Entity implements \JsonSerializable{
         try{
             if($this->isNew()){
                 $this->checkPermission('create');
-
+                $is_new = true;
 
                 if($this->usesOriginSubsite()){
                     $this->_subsiteId = $app->getCurrentSubsiteId();
@@ -540,6 +551,8 @@ abstract class Entity implements \JsonSerializable{
 
             }else{
                 $this->checkPermission('modify');
+                $is_new = false;
+
             }
             $app->em->persist($this);
 
@@ -560,13 +573,19 @@ abstract class Entity implements \JsonSerializable{
                     $app->em->flush();
                 }
             }
+            $this->refresh();
+            if($is_new){
+                $this->_newCreatedRevision();
+            } else {
+                $this->_newModifiedRevision();
+            }
 
             // delete the entity cache
             $repo = $this->repo();
             if($repo->usesCache()){
                 $repo->deleteEntityCache($this->id);
             }
-
+            
         }catch(Exceptions\PermissionDenied $e){
             if(!$requests)
                 throw $e;
@@ -708,7 +727,7 @@ abstract class Entity implements \JsonSerializable{
         if(!method_exists($class, 'getValidations')) {
             return $errors;
         }
-        
+
         foreach($class::getValidations() as $property => $validations){
 
             if(!$this->$property && !key_exists('required', $validations))
@@ -719,7 +738,6 @@ abstract class Entity implements \JsonSerializable{
                 $validation = trim($validation);
 
                 $ok = true;
-
 
                 if($validation == 'required'){
                     if (is_string($this->$property)) {
@@ -790,6 +808,14 @@ abstract class Entity implements \JsonSerializable{
 
         $app->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').insert:before', $args);
         $app->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').save:before', $args);
+        
+        
+        if($this->usesPermissionCache()){
+            if($this->usesAgentRelation()){
+                $this->deleteUsersWithControlCache();
+            }
+            $this->addToRecreatePermissionsCacheList();
+        }
     }
 
     /**
@@ -837,6 +863,14 @@ abstract class Entity implements \JsonSerializable{
         $app = App::i();
 
         $app->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').remove:before', $args);
+        
+        
+        if($this->usesPermissionCache()){
+            if($this->usesAgentRelation()){
+                $this->deleteUsersWithControlCache();
+            }
+            $this->addToRecreatePermissionsCacheList();
+        }
     }
 
     /**
@@ -857,6 +891,14 @@ abstract class Entity implements \JsonSerializable{
 
 
         $app->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').remove:after', $args);
+
+        if($this->usesRevision()) {
+            //$this->_newDeletedRevision();
+        }
+        
+        if($this->usesPermissionCache()){
+            $this->deletePermissionsCache();
+        }
     }
 
     /**
@@ -878,22 +920,32 @@ abstract class Entity implements \JsonSerializable{
         $app->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').save:before', $args);
 
         if (property_exists($this, 'updateTimestamp')) {
+            return;
             $this->updateTimestamp = new \DateTime;
             if($this->sentNotification){
                 $entity = $this;
                 $nid = $this->sentNotification;
                 $app->hook('entity(' . $hook_class_path . ').update:after', function() use($app, $entity, $nid) {
-                    if($this->equals($entity)){
+                    if($this->equals($entity) && $app->user->equals($this->getOwnerUser())){
                         $app->log->debug("notification id: $nid");
                         $notification = $app->repo('Notification')->find($nid);
-                        $notification->delete();
-                        $this->sentNotification = 0;
-                        $this->save();
+                        if($notification){
+                            $notification->delete();
+                            $this->sentNotification = 0;
+                            $this->save();
 
-                        $app->em->flush();
+                            $app->em->flush();
+                        }
                     }
                 });
             }
+        }
+        
+        if($this->usesPermissionCache()){
+            if($this->usesAgentRelation()){
+                $this->deleteUsersWithControlCache();
+            }
+            $this->addToRecreatePermissionsCacheList();
         }
     }
 
@@ -922,7 +974,5 @@ abstract class Entity implements \JsonSerializable{
 
         $app->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').update:after', $args);
         $app->applyHookBoundTo($this, 'entity(' . $hook_class_path . ').save:after', $args);
-
     }
-
 }
